+++ docs/RELEASE-NOTES.rst
@@ -6,10 +6,15 @@ Next Release
 
 Added Functionality
 ```````````````````
-* Added Support for AS3 3.38.0
-* CRD:
+
+**What’s new:**
+    * Next generation routes
+        * Added support for namespaceLabel in global extended ConfigMap
+    * CRD:
         * AllowSourceRange support for VirtualServer CRs and Policy CRs. See `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/customResource/>`_
-* Added support to configure netmask for Virtual Server for Ingress. See `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/ingress/>`_
+    * Added support to configure netmask for Virtual Server for Ingress. See `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/ingress/>`_
+    * Added Support for AS3 3.38.0
+
 Bug Fixes
 ````````````
 * :issues:`2361` Allow monitoring of an alias port in VirtualServer and TransportServer